fix: CI failures — file size, clippy, compile errors (#62)
- Add #[allow(clippy::too_many_arguments)] to producer_task and consumer_task
- Extract FlvStreamNormalizer to flv_normalizer.rs (endpoint_task.rs: 1121 -> 997 lines)
- Extract run_rescue_loop() to rescue.rs (replaces 100-line inline rescue block)
- Extract FlvStreamNormalizer tests + write-failure test to endpoint_task_flv_tests.rs (endpoint_task_tests.rs: 1140 -> 997 lines)
- Fix rescue_tests.rs: use super::* (not super::rescue::*)
- Add missing delivery_mode/rescue_eta_secs fields to EndpointStats literal in test_stats_struct_serializes
